DynaFed note for con_max_block_sig_size
Included a note that the signature size constant under DynaFed is 77, rather than 74 prior to DynaFed. The formula for con_max_block_sig_size is affected. The guide doesnt yet use DynaFed so this may not be the most appropriate place for this note, feel free to reject it and put in a more appropriate location.